changed the transactions classes to inherit from DataCreatorDTO

+++ Thandizo.DataModels/Patients/PatientDailyStatusDTO.cs
@@ -5,7 +5,7 @@ namespace Thandizo.DataModels.Patients
 {
     public class PatientDailyStatusDTO : DataCreatorDTO
     {
-        public int SubmissionId { get; set; }
+        public long SubmissionId { get; set; }
         public int SymptomId { get; set; }
         public long PatientId { get; set; }
         public DateTime DateSubmitted { get; set; }
