+++ Thandizo.DataModels/Patients/PatientTravelHistoryDTO.cs
@@ -4,7 +4,7 @@ namespace Thandizo.DataModels.Patients
 {
     public class PatientTravelHistoryDTO : DataCreatorDTO
     {
-        public int TravelId { get; set; }
+        public long TravelId { get; set; }
         public string CountryCode { get; set; }
         public long PatientId { get; set; }
     }
